refactor: move scrubber into its own package
Also part of https://github.com/ooni/probe/issues/1687
