+++ internal/scrubber/scrubber.go
@@ -1,4 +1,4 @@
-package errorx
+package scrubber
 
 import "regexp"
 
@@ -64,7 +64,7 @@ func scrub(b []byte) []byte {
 }
 
 // Scrub sanitizes a string containing an error such that
-// any occurrence of IP endpoints is scrubbed
+// any occurrence of IP endpoints is scrubbed.
 func Scrub(s string) string {
 	return string(scrub([]byte(s)))
 }
