+++ internal/scrubber/scrubber_test.go
@@ -1,4 +1,4 @@
-package errorx
+package scrubber
 
 import (
 	"testing"
